repr changed to show error instead of variance
